Fixed various circular import errors by moving code around.
