+++ db/tests/test_tables.py
@@ -1,11 +1,13 @@
-import decimal
 from unittest.mock import call, patch
 import pytest
 from sqlalchemy import MetaData, select, Column, String, Table, ForeignKey, Integer
 from sqlalchemy.exc import NoSuchTableError
 from psycopg2.errors import DependentObjectsStillExist
+
 from db import tables, constants, columns
+from db.types import inference
 from db.tests.types import fixtures
+
 
 ROSTER = "Roster"
 TEACHERS = "Teachers"
@@ -14,6 +16,7 @@ EXTRACTED_COLS = ["Teacher", "Teacher Email"]
 REM_MOVE_COL = ["Subject"]
 REM_MOVE_COLS = ["Student Name", "Student Email"]
 FKEY_COL = f"{TEACHERS}_{constants.ID}"
+
 
 # We need to set these variables when the file loads, or pytest can't
 # properly detect the fixtures.  Importing them directly results in a
@@ -461,75 +464,13 @@ def test_infer_table_column_types_doesnt_touch_defaults(engine_with_schema):
     tables.create_mathesar_table(
         table_name, schema, column_list, engine
     )
-    with patch.object(tables.inference, "infer_column_type") as mock_infer:
+    with patch.object(inference, "infer_column_type") as mock_infer:
         tables.update_table_column_types(
             schema,
             table_name,
             engine
         )
     mock_infer.assert_not_called()
-
-
-def test_get_column_cast_records(engine_email_type):
-    COL1 = "col1"
-    COL2 = "col2"
-    col1 = Column(COL1, String)
-    col2 = Column(COL2, String)
-    column_list = [col1, col2]
-    engine, schema = engine_email_type
-    table_name = "table_with_columns"
-    table = tables.create_mathesar_table(
-        table_name, schema, column_list, engine
-    )
-    ins = table.insert().values(
-        [{COL1: 'one', COL2: 1}, {COL1: 'two', COL2: 2}]
-    )
-    with engine.begin() as conn:
-        conn.execute(ins)
-    COL1_MOD = COL1 + "_mod"
-    COL2_MOD = COL2 + "_mod"
-    column_definitions = [
-        {"name": "mathesar_id", "type": "INTEGER"},
-        {"name": COL1_MOD, "type": "VARCHAR"},
-        {"name": COL2_MOD, "type": "NUMERIC"},
-    ]
-    records = tables.get_column_cast_records(engine, table, column_definitions)
-    for record in records:
-        assert (
-            type(record[COL1 + "_mod"]) == str
-            and type(record[COL2 + "_mod"]) == decimal.Decimal
-        )
-
-
-def test_get_column_cast_records_options(engine_email_type):
-    COL1 = "col1"
-    COL2 = "col2"
-    col1 = Column(COL1, String)
-    col2 = Column(COL2, String)
-    column_list = [col1, col2]
-    engine, schema = engine_email_type
-    table_name = "table_with_columns"
-    table = tables.create_mathesar_table(
-        table_name, schema, column_list, engine
-    )
-    ins = table.insert().values(
-        [{COL1: 'one', COL2: 1}, {COL1: 'two', COL2: 2}]
-    )
-    with engine.begin() as conn:
-        conn.execute(ins)
-    COL1_MOD = COL1 + "_mod"
-    COL2_MOD = COL2 + "_mod"
-    column_definitions = [
-        {"name": "mathesar_id", "type": "INTEGER"},
-        {"name": COL1_MOD, "type": "VARCHAR"},
-        {"name": COL2_MOD, "type": "NUMERIC", "type_options": {"precision": 5, "scale": 2}},
-    ]
-    records = tables.get_column_cast_records(engine, table, column_definitions)
-    for record in records:
-        assert (
-            type(record[COL1 + "_mod"]) == str
-            and type(record[COL2 + "_mod"]) == decimal.Decimal
-        )
 
 
 def test_update_table_column_types_infers_non_default_types(engine_with_schema):
@@ -541,7 +482,7 @@ def test_update_table_column_types_infers_non_default_types(engine_with_schema):
     tables.create_mathesar_table(
         table_name, schema, column_list, engine
     )
-    with patch.object(tables.inference, "infer_column_type") as mock_infer:
+    with patch.object(inference, "infer_column_type") as mock_infer:
         tables.update_table_column_types(
             schema,
             table_name,
@@ -571,7 +512,7 @@ def test_update_table_column_types_skips_pkey_columns(engine_with_schema):
     tables.create_mathesar_table(
         table_name, schema, column_list, engine
     )
-    with patch.object(tables.inference, "infer_column_type") as mock_infer:
+    with patch.object(inference, "infer_column_type") as mock_infer:
         tables.update_table_column_types(
             schema,
             table_name,
@@ -584,7 +525,7 @@ def test_update_table_column_types_skips_fkey_columns(
         extracted_remainder_roster
 ):
     _, remainder, _, engine, schema = extracted_remainder_roster
-    with patch.object(tables.inference, "infer_column_type") as mock_infer:
+    with patch.object(inference, "infer_column_type") as mock_infer:
         tables.update_table_column_types(
             schema,
             remainder.name,
