+++ db/tests/types/test_inference.py
@@ -78,7 +78,7 @@ def test_table_inference(engine_email_type, type_, value_list, expect_type):
         results = conn.execute(select(input_table))
     original_table = results.fetchall()
 
-    inferred_types = tables.infer_table_column_types(
+    inferred_types = inference.infer_table_column_types(
         schema,
         TEST_TABLE,
         engine
@@ -104,8 +104,8 @@ def test_table_inference_drop_temp(engine_email_type):
     with patch.object(tables.inference, "infer_column_type") as mock_infer:
         mock_infer.side_effect = Exception()
         with pytest.raises(Exception):
-            tables.infer_table_column_types(schema, TEST_TABLE, engine)
-    tables.infer_table_column_types(schema, TEST_TABLE, engine)
+            inference.infer_table_column_types(schema, TEST_TABLE, engine)
+    inference.infer_table_column_types(schema, TEST_TABLE, engine)
 
 
 def test_table_inference_same_name(engine_email_type):
@@ -118,7 +118,7 @@ def test_table_inference_same_name(engine_email_type):
     with engine.begin() as conn:
         results = conn.execute(select(table))
     original_table = results.fetchall()
-    tables.infer_table_column_types(schema, TEST_TABLE, engine)
+    inference.infer_table_column_types(schema, TEST_TABLE, engine)
     with engine.begin() as conn:
         results = conn.execute(select(table))
     new_table = results.fetchall()
