+++ db/types/__init__.py
@@ -1,11 +1,11 @@
 from sqlalchemy import DECIMAL as sa_decimal
 from db.types import email
-from db.types.alteration import DECIMAL
+from db.types.base import PostgresType
 
 
 CUSTOM_TYPE_DICT = {
     # For some reason, SQLAlchemy doesn't add DECIMAL to the default
     # ischema_names supported by a PostgreSQL engine
-    DECIMAL: sa_decimal,
+    PostgresType.DECIMAL.value: sa_decimal,
     email.DB_TYPE: email.Email,
 }
