+++ mathesar/utils/tables.py
@@ -1,9 +1,10 @@
 from sqlalchemy import MetaData
 
+from db.tables import create_mathesar_table, get_oid_from_table
+from db.types.inference import infer_table_column_types
 from mathesar.database.base import create_mathesar_engine
 from mathesar.imports.csv import create_table_from_csv
 from mathesar.models import Table
-from db.tables import infer_table_column_types, create_mathesar_table, get_oid_from_table
 
 
 TABLE_NAME_TEMPLATE = 'Table'
